refactor: move fingerprint to activity
