+++ app/src/main/java/com/payone/pcpclientandroiddemo/MainActivity.kt
@@ -1,27 +1,11 @@
 package com.payone.pcpclientandroiddemo
 
-import android.app.Activity
+
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.widget.Button
-import android.widget.TextView
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.platform.ComposeView
-import com.google.android.gms.wallet.AutoResolveHelper
-import com.google.android.gms.wallet.PaymentsClient
-import com.google.android.gms.wallet.Wallet
-import com.google.android.gms.wallet.WalletConstants
-import com.google.android.gms.wallet.PaymentDataRequest
-import com.payone.pcpclientandroiddemo.fingerprint.FingerprintTokenizerHandler
-import com.payone.pcpclientandroiddemo.cctokenizer.CCTokenizerHandler
-import com.paypal.android.paymentbuttons.PayPalButton
-import com.payone.pcpclientandroiddemo.gpay.GooglePayHandler
-import com.payone.pcpclientandroiddemo.gpay.GooglePayRequestJson
-import com.payone.pcpclientandroiddemo.paypal.PayPalHandler
+
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
